requirements.txt
google-genai==1.73.1
openpyxl==3.1.5
pandas==3.0.2
pdfplumber==0.11.9
pydantic==2.13.3
pytest==9.0.3
python-dotenv==1.2.2
streamlit==1.56.0
xlsxwriter==3.2.9

